refactor(modifiers): split defense authority
